Testing new way of fetching after login
